+++ src/hooks/useAllPolicyExpenseChatReportActions.ts
@@ -1,6 +1,5 @@
-import {isPolicyExpenseChat, isThread} from '@libs/ReportUtils';
+import {getAllPolicyExpenseChatReportActions} from '@libs/ReportUtils';
 import ONYXKEYS from '@src/ONYXKEYS';
-import type {ReportActions} from '@src/types/onyx';
 import useOnyx from './useOnyx';
 
 /**
@@ -10,19 +9,7 @@ function useAllPolicyExpenseChatReportActions() {
     const [allReports] = useOnyx(ONYXKEYS.COLLECTION.REPORT);
     const [allReportActions] = useOnyx(ONYXKEYS.COLLECTION.REPORT_ACTIONS);
 
-    const filteredReportActions: Record<string, ReportActions> = {};
-    for (const report of Object.values(allReports ?? {})) {
-        if (!report?.reportID || !isPolicyExpenseChat(report) || isThread(report)) {
-            continue;
-        }
-        const key = `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${report.reportID}`;
-        const actions = allReportActions?.[key];
-        if (actions) {
-            filteredReportActions[key] = actions;
-        }
-    }
-
-    return filteredReportActions;
+    return getAllPolicyExpenseChatReportActions(allReports, allReportActions);
 }
 
 export default useAllPolicyExpenseChatReportActions;
