+++ src/libs/ReportUtils.ts
@@ -1798,6 +1798,25 @@ function isThread(report: OnyxInputOrEntry<Report>): report is Thread {
 }
 
 /**
+ * Returns reportActions filtered to only policy expense chat reports (non-thread).
+ */
+function getAllPolicyExpenseChatReportActions(allReports: OnyxCollection<Report>, allReportActions: OnyxCollection<ReportActions>) {
+    const filteredReportActions: Record<string, ReportActions> = {};
+    for (const report of Object.values(allReports ?? {})) {
+        if (!report?.reportID || !isPolicyExpenseChat(report) || isThread(report)) {
+            continue;
+        }
+        const key = `${ONYXKEYS.COLLECTION.REPORT_ACTIONS}${report.reportID}`;
+        const actions = allReportActions?.[key];
+        if (actions) {
+            filteredReportActions[key] = actions;
+        }
+    }
+
+    return filteredReportActions;
+}
+
+/**
  * Returns true if report is of type chat and has a parent and is therefore a Thread.
  */
 function isChatThread(report: OnyxInputOrEntry<Report>): report is Thread {
@@ -13004,6 +13023,7 @@ function shouldShowMarkAsDone({isTrackIntentUser, report, policy}: {isTrackInten
 
 export {
     areAllRequestsBeingSmartScanned,
+    getAllPolicyExpenseChatReportActions,
     buildConciergeGreetingReportAction,
     buildOptimisticAddCommentReportAction,
     buildOptimisticApprovedReportAction,
